Move squerryl types to squerryl package, leave query traits in api package
