+++ db/src/main/scala/squeryl/tables/BlockBondsTable.scala
@@ -1,4 +1,4 @@
-package sdk.db
+package squeryl.tables
 
 import sdk.api.data.*
 
@@ -8,9 +8,4 @@ case class BlockBondsTable(blockId: Long, bondId: Long)
 object BlockBondsTable {
   def toDb(blockBonds: BlockBonds): BlockBondsTable   = BlockBondsTable(blockBonds.blockId, blockBonds.bondId)
   def fromDb(blockBonds: BlockBondsTable): BlockBonds = BlockBonds(blockBonds.blockId, blockBonds.bondId)
-}
-
-trait BlockBondsDbApi[F[_]] {
-  def insert(blockBonds: BlockBonds): F[BlockBonds]
-  def getByBlock(blockId: Long): F[Seq[BlockBonds]]
 }
