+++ db/src/main/scala/squeryl/tables/BlockDeploysTable.scala
@@ -1,14 +1,9 @@
-package sdk.db
+package squeryl.tables
 
 import sdk.api.data.*
 
 @SuppressWarnings(Array("org.wartremover.warts.FinalCaseClass"))
 case class BlockDeploysTable(blockId: Long, deployId: Long)
-
-trait BlockDeploysDbApi[F[_]] {
-  def insert(blockDeploys: BlockDeploys): F[BlockDeploys]
-  def getByBlock(blockId: Long): F[Seq[BlockDeploys]]
-}
 
 object BlockDeploysTable {
   def toDb(blockDeploys: BlockDeploys): BlockDeploysTable =
