+++ db/src/main/scala/squeryl/tables/BlockTable.scala
@@ -1,6 +1,5 @@
-package sdk.db
+package squeryl.tables
 
-import sdk.DbTable
 import sdk.api.data.*
 
 @SuppressWarnings(Array("org.wartremover.warts.FinalCaseClass"))
@@ -72,12 +71,4 @@ object BlockTable {
     block.signature,
     block.status,
   )
-}
-
-trait BlockDbApi[F[_]] {
-  def insert(block: Block, senderId: Long): F[Long]
-  def update(id: Long, block: Block, senderId: Long): F[Unit]
-
-  def getById(id: Long): F[Option[Block]]
-  def getByHash(hash: Array[Byte]): F[Option[Block]]
 }
