+++ db/src/main/scala/squeryl/tables/BondTable.scala
@@ -1,6 +1,5 @@
-package sdk.db
+package squeryl.tables
 
-import sdk.DbTable
 import sdk.api.data.Bond
 
 @SuppressWarnings(Array("org.wartremover.warts.FinalCaseClass"))
@@ -13,11 +12,4 @@ case class BondTable(
 object BondTable {
   def toDb(id: Long, bond: Bond, validatorId: Long): BondTable = BondTable(id, validatorId, bond.stake)
   def fromDb(bond: BondTable, validator: ValidatorTable): Bond = Bond(ValidatorTable.fromDb(validator), bond.stake)
-}
-
-trait BondDbApi[F[_]] {
-  def insert(bond: Bond, validatorId: Long): F[Long]
-  def update(id: Long, bond: Bond, validatorId: Long): F[Unit]
-
-  def getById(id: Long): F[Option[Bond]]
 }
