+++ db/src/main/scala/squeryl/tables/DbTable.scala
@@ -1,0 +1,7 @@
+package squeryl.tables
+
+import org.squeryl.KeyedEntity
+
+trait DbTable extends KeyedEntity[Long] {
+  def name: String = this.getClass.getSimpleName.takeWhile(_ != '$').replace("Table", "")
+}
