+++ db/src/main/scala/squeryl/tables/DeployTable.scala
@@ -1,6 +1,5 @@
-package sdk.db
+package squeryl.tables
 
-import sdk.DbTable
 import sdk.api.data.Deploy
 
 @SuppressWarnings(Array("org.wartremover.warts.FinalCaseClass"))
@@ -39,12 +38,4 @@ object DeployTable {
     deploy.timestamp,
     deploy.validAfterBlockNumber,
   )
-}
-
-trait DeployDbApi[F[_]] {
-  def insert(deploy: Deploy): F[Long]
-  def update(id: Long, deploy: Deploy): F[Unit]
-
-  def getById(id: Long): F[Option[Deploy]]
-  def getByHash(hash: Array[Byte]): F[Option[Deploy]]
 }
