+++ sdk/src/main/scala/sdk/api/BlockBondsDbApi.scala
@@ -1,0 +1,8 @@
+package sdk.api
+
+import sdk.api.data.BlockBonds
+
+trait BlockBondsDbApi[F[_]] {
+  def insert(blockBonds: BlockBonds): F[BlockBonds]
+  def getByBlock(blockId: Long): F[Seq[BlockBonds]]
+}
