+++ sdk/src/main/scala/sdk/api/BlockDeploysDbApi.scala
@@ -1,0 +1,8 @@
+package sdk.api
+
+import sdk.api.data.BlockDeploys
+
+trait BlockDeploysDbApi[F[_]] {
+  def insert(blockDeploys: BlockDeploys): F[BlockDeploys]
+  def getByBlock(blockId: Long): F[Seq[BlockDeploys]]
+}
