+++ sdk/src/main/scala/sdk/api/BlockJustificationsDbApi.scala
@@ -1,0 +1,8 @@
+package sdk.api
+
+import sdk.api.data.BlockJustifications
+
+trait BlockJustificationsDbApi[F[_]] {
+  def insert(blockJustifications: BlockJustifications): F[BlockJustifications]
+  def getByBlock(latestBlockId: Long): F[Seq[BlockJustifications]]
+}
